+++ react/package.json
@@ -1,40 +1,40 @@
 {
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject"
-  },
-  "dependencies": {
-    "@react-native-community/datetimepicker": "3.0.0",
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-navigation/bottom-tabs": "^5.9.2",
-    "@react-navigation/material-top-tabs": "^5.2.19",
-    "@react-navigation/native": "^5.7.6",
-    "@react-navigation/stack": "^5.4.1",
-    "expo": "^39.0.3",
-    "formik": "^2.1.5",
-    "package": "^1.0.1",
-    "react": "~16.9.0",
-    "react-dom": "~16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
-    "react-native-gesture-handler": "~1.7.0",
-    "react-native-reanimated": "~1.13.0",
-    "react-native-safe-area-context": "3.1.4",
-    "react-native-screens": "~2.10.1",
-    "react-native-svg": "12.1.0",
-    "react-native-tab-view": "^2.15.2",
-    "react-native-web": "~0.11.7",
-    "react-navigation": "^4.3.9",
-    "react-navigation-stack": "^2.5.1",
-    "tailwind-rn": "^1.1.0",
-    "yup": "^0.29.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.8.6",
-    "babel-preset-expo": "~8.1.0"
-  },
-  "private": true
+	"main": "node_modules/expo/AppEntry.js",
+	"scripts": {
+		"start": "expo start",
+		"android": "expo start --android",
+		"ios": "expo start --ios",
+		"web": "expo start --web",
+		"eject": "expo eject"
+	},
+	"dependencies": {
+		"@react-native-community/datetimepicker": "3.0.0",
+		"@react-native-community/masked-view": "0.1.10",
+		"@react-navigation/bottom-tabs": "^5.9.2",
+		"@react-navigation/material-top-tabs": "^5.2.19",
+		"@react-navigation/native": "^5.7.6",
+		"@react-navigation/stack": "^5.4.1",
+		"expo": "39.0.0",
+		"formik": "^2.1.5",
+		"package": "^1.0.1",
+		"react": "~16.9.0",
+		"react-dom": "~16.9.0",
+		"react-native": "https://github.com/expo/react-native/archive/sdk-39.0.1.tar.gz",
+		"react-native-gesture-handler": "~1.7.0",
+		"react-native-reanimated": "~1.13.0",
+		"react-native-safe-area-context": "3.1.4",
+		"react-native-screens": "~2.10.1",
+		"react-native-svg": "12.1.0",
+		"react-native-tab-view": "^2.15.2",
+		"react-native-web": "~0.11.7",
+		"react-navigation": "^4.3.9",
+		"react-navigation-stack": "^2.5.1",
+		"tailwind-rn": "^1.1.0",
+		"yup": "^0.29.2"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.8.6",
+		"babel-preset-expo": "~8.1.0"
+	},
+	"private": true
 }
