+++ react/src/comum/assets/IconeAdicionar.js
