+++ react/src/comum/components/IndicadorNavegacao.js
